Update main.cpp
Added signal handler that will send a kill signal to all scripts opened by main process. Updated button handling in the following way:
If a button on the main hand is pressed, the drawing will update.
If a button on the main hand is released, the drawing will remain.
If a button on the mod hand is pressed, the drawing will update.
If a button on the mod hand is released, the drawing will update.
While the signal handler has been tested, I have not yet tested the button handling due to time constraints.